fix: add configurable security options for trivy scans
