+++ .env.example
@@ -40,6 +40,11 @@ ALLOW_DOWNGRADE=false
 # Docker Configuration
 # DOCKER_HOST=unix:///var/run/docker.sock  # Default: direct socket access
 # DOCKER_HOST=tcp://docker-socket-proxy:2375  # Example: via socket proxy for enhanced security
+# Trivy scan container security options (optional, useful on SELinux hosts)
+# Separate multiple entries with commas or new lines, for example: label=disable,label=type:container_runtime_t
+# TRIVY_SECURITY_OPTS=label=disable
+# Run Trivy scan containers in privileged mode only when security options alone are not enough
+# TRIVY_PRIVILEGED=false
 
 # Agent Configuration (optional)
 # AGENT_MODE=true
