+++ backend/internal/services/vulnerability_service.go
@@ -1272,21 +1272,8 @@ func (s *VulnerabilityService) createBatchTrivyContainer(ctx context.Context, tr
 		},
 	}
 
-	hostConfig := &containertypes.HostConfig{
-		NetworkMode: containertypes.NetworkMode(s.getTrivyNetworkModeInternal()),
-		Mounts: []mounttypes.Mount{
-			{
-				Type:   mounttypes.TypeBind,
-				Source: "/var/run/docker.sock",
-				Target: "/var/run/docker.sock",
-			},
-			{
-				Type:   mounttypes.TypeVolume,
-				Source: cacheVolume,
-				Target: trivyCacheMountTarget,
-			},
-		},
-	}
+	securityOpts, privileged := s.getTrivyRuntimeSecurityInternal()
+	hostConfig := buildTrivyBatchHostConfigInternal(cacheVolume, s.getTrivyNetworkModeInternal(), securityOpts, privileged)
 
 	resources, cpuSet, applyLimits := s.getTrivyContainerResourceOptionsInternal(ctx, dockerClient)
 	applyTrivyContainerResourcesInternal(hostConfig, resources, cpuSet, applyLimits)
@@ -1609,6 +1596,34 @@ func normalizeTrivyNetworkModeInternal(networkMode string) string {
 	return networkMode
 }
 
+func parseTrivySecurityOptsInternal(value string) []string {
+	value = strings.TrimSpace(value)
+	if value == "" {
+		return nil
+	}
+
+	value = strings.NewReplacer("\r\n", "\n", "\r", "\n").Replace(value)
+	parts := strings.FieldsFunc(value, func(r rune) bool {
+		return r == ',' || r == '\n'
+	})
+	if len(parts) == 0 {
+		return nil
+	}
+
+	opts := make([]string, 0, len(parts))
+	for _, part := range parts {
+		if opt := strings.TrimSpace(part); opt != "" {
+			opts = append(opts, opt)
+		}
+	}
+
+	if len(opts) == 0 {
+		return nil
+	}
+
+	return opts
+}
+
 func (s *VulnerabilityService) getTrivyNetworkModeInternal() string {
 	if s.settingsService == nil {
 		return DefaultTrivyNetworkMode
@@ -1620,6 +1635,19 @@ func (s *VulnerabilityService) getTrivyNetworkModeInternal() string {
 	}
 
 	return normalizeTrivyNetworkModeInternal(cfg.TrivyNetwork.Value)
+}
+
+func (s *VulnerabilityService) getTrivyRuntimeSecurityInternal() ([]string, bool) {
+	if s.settingsService == nil {
+		return nil, false
+	}
+
+	cfg := s.settingsService.GetSettingsConfig()
+	if cfg == nil {
+		return nil, false
+	}
+
+	return parseTrivySecurityOptsInternal(cfg.TrivySecurityOpts.Value), cfg.TrivyPrivileged.IsTrue()
 }
 
 func (s *VulnerabilityService) getTrivyScanTimeoutArg() string {
@@ -1806,7 +1834,37 @@ func buildTrivyContainerConfig(trivyImage string, cmdArgs []string) *containerty
 	}
 }
 
-func buildTrivyHostConfig(cacheVolume string, tempFiles []string, resources containertypes.Resources, cpuSet string, applyLimits bool, networkMode string) *containertypes.HostConfig {
+func buildTrivyBatchHostConfigInternal(cacheVolume, networkMode string, securityOpts []string, privileged bool) *containertypes.HostConfig {
+	hostConfig := &containertypes.HostConfig{
+		NetworkMode: containertypes.NetworkMode(normalizeTrivyNetworkModeInternal(networkMode)),
+		Mounts: []mounttypes.Mount{
+			{
+				Type:   mounttypes.TypeBind,
+				Source: "/var/run/docker.sock",
+				Target: "/var/run/docker.sock",
+			},
+			{
+				Type:   mounttypes.TypeVolume,
+				Source: cacheVolume,
+				Target: trivyCacheMountTarget,
+			},
+		},
+	}
+
+	applyTrivyRuntimeSecurityInternal(hostConfig, securityOpts, privileged)
+	return hostConfig
+}
+
+func buildTrivyHostConfig(
+	cacheVolume string,
+	tempFiles []string,
+	resources containertypes.Resources,
+	cpuSet string,
+	applyLimits bool,
+	networkMode string,
+	securityOpts []string,
+	privileged bool,
+) *containertypes.HostConfig {
 	hostConfig := &containertypes.HostConfig{
 		NetworkMode: containertypes.NetworkMode(normalizeTrivyNetworkModeInternal(networkMode)),
 		// Keep single-scan containers until explicit cleanup so we can reliably
@@ -1826,10 +1884,22 @@ func buildTrivyHostConfig(cacheVolume string, tempFiles []string, resources cont
 		},
 	}
 
+	applyTrivyRuntimeSecurityInternal(hostConfig, securityOpts, privileged)
 	applyTrivyContainerResourcesInternal(hostConfig, resources, cpuSet, applyLimits)
 
 	addTrivyTempFileMounts(hostConfig, tempFiles)
 	return hostConfig
+}
+
+func applyTrivyRuntimeSecurityInternal(hostConfig *containertypes.HostConfig, securityOpts []string, privileged bool) {
+	if hostConfig == nil {
+		return
+	}
+
+	if len(securityOpts) > 0 {
+		hostConfig.SecurityOpt = append([]string(nil), securityOpts...)
+	}
+	hostConfig.Privileged = privileged
 }
 
 func applyTrivyContainerResourcesInternal(hostConfig *containertypes.HostConfig, resources containertypes.Resources, cpuSet string, applyLimits bool) {
@@ -1938,11 +2008,15 @@ func logTrivyContainerStartupRequestInternal(ctx context.Context, scope string, 
 	autoRemove := false
 	mounts := []string{}
 	networkMode := ""
+	securityOpts := []string{}
+	privileged := false
 	if hostConfig != nil {
 		resources = hostConfig.Resources
 		autoRemove = hostConfig.AutoRemove
 		mounts = trivyMountSummariesInternal(hostConfig)
 		networkMode = string(hostConfig.NetworkMode)
+		securityOpts = append([]string(nil), hostConfig.SecurityOpt...)
+		privileged = hostConfig.Privileged
 	}
 
 	slog.DebugContext(ctx,
@@ -1952,6 +2026,8 @@ func logTrivyContainerStartupRequestInternal(ctx context.Context, scope string, 
 		"cmd", config.Cmd,
 		"entrypoint", config.Entrypoint,
 		"networkMode", networkMode,
+		"securityOpts", securityOpts,
+		"privileged", privileged,
 		"autoRemove", autoRemove,
 		"mounts", mounts,
 		"nanoCPUs", resources.NanoCPUs,
@@ -2784,7 +2860,17 @@ func (s *VulnerabilityService) runTrivyScan(ctx context.Context, trivyImage stri
 
 	config := buildTrivyContainerConfig(trivyImage, cmdArgs)
 	resources, cpuSet, applyLimits := s.getTrivyContainerResourceOptionsInternal(ctx, dockerClient)
-	hostConfig := buildTrivyHostConfig(cacheVolume, tempFiles, resources, cpuSet, applyLimits, s.getTrivyNetworkModeInternal())
+	securityOpts, privileged := s.getTrivyRuntimeSecurityInternal()
+	hostConfig := buildTrivyHostConfig(
+		cacheVolume,
+		tempFiles,
+		resources,
+		cpuSet,
+		applyLimits,
+		s.getTrivyNetworkModeInternal(),
+		securityOpts,
+		privileged,
+	)
 	s.setScanPhaseInternal(imageID, vulnerability.ScanPhaseCreatingContainer)
 
 	stdoutFile, stderrFile, duration, statusCode, containerID, err := s.runTrivyContainer(ctx, dockerClient, imageID, config, hostConfig)
