+++ backend/internal/services/vulnerability_service_test.go
@@ -296,11 +296,13 @@ func TestBuildTrivyHostConfig_ExcludesResourcesWhenLimitsDisabled(t *testing.T) 
 		MemorySwap: int64(1_073_741_824),
 	}
 
-	hostConfig := buildTrivyHostConfig("cache-volume", nil, resources, "", false, "bridge")
+	hostConfig := buildTrivyHostConfig("cache-volume", nil, resources, "", false, "bridge", nil, false)
 	require.Equal(t, containertypes.NetworkMode("bridge"), hostConfig.NetworkMode)
 	require.Equal(t, int64(0), hostConfig.Resources.NanoCPUs)
 	require.Equal(t, int64(0), hostConfig.Resources.Memory)
 	require.Equal(t, int64(0), hostConfig.Resources.MemorySwap)
+	require.Empty(t, hostConfig.SecurityOpt)
+	require.False(t, hostConfig.Privileged)
 }
 
 func TestBuildTrivyHostConfig_IncludesResourcesWhenLimitsEnabled(t *testing.T) {
@@ -310,7 +312,7 @@ func TestBuildTrivyHostConfig_IncludesResourcesWhenLimitsEnabled(t *testing.T) {
 		MemorySwap: int64(1_073_741_824),
 	}
 
-	hostConfig := buildTrivyHostConfig("cache-volume", nil, resources, "", true, "arcane-external")
+	hostConfig := buildTrivyHostConfig("cache-volume", nil, resources, "", true, "arcane-external", nil, false)
 	require.Equal(t, containertypes.NetworkMode("arcane-external"), hostConfig.NetworkMode)
 	require.Equal(t, resources.NanoCPUs, hostConfig.Resources.NanoCPUs)
 	require.Equal(t, resources.Memory, hostConfig.Resources.Memory)
@@ -324,12 +326,72 @@ func TestBuildTrivyHostConfig_UsesCPUSetAndClearsNanoCPUs(t *testing.T) {
 		MemorySwap: int64(1_073_741_824),
 	}
 
-	hostConfig := buildTrivyHostConfig("cache-volume", nil, resources, "0-1", true, "bridge")
+	hostConfig := buildTrivyHostConfig("cache-volume", nil, resources, "0-1", true, "bridge", nil, false)
 	require.Equal(t, containertypes.NetworkMode("bridge"), hostConfig.NetworkMode)
 	require.Equal(t, int64(0), hostConfig.Resources.NanoCPUs)
 	require.Equal(t, "0-1", hostConfig.Resources.CpusetCpus)
 	require.Equal(t, resources.Memory, hostConfig.Resources.Memory)
 	require.Equal(t, resources.MemorySwap, hostConfig.Resources.MemorySwap)
+}
+
+func TestParseTrivySecurityOptsInternal(t *testing.T) {
+	tests := []struct {
+		name  string
+		value string
+		want  []string
+	}{
+		{name: "empty", value: "", want: nil},
+		{name: "whitespace only", value: " \n\t ", want: nil},
+		{name: "single", value: "label=disable", want: []string{"label=disable"}},
+		{name: "comma separated", value: "label=disable,label=type:container_runtime_t", want: []string{"label=disable", "label=type:container_runtime_t"}},
+		{name: "newline separated", value: "label=disable\nlabel=type:container_runtime_t", want: []string{"label=disable", "label=type:container_runtime_t"}},
+		{name: "mixed whitespace", value: " label=disable,\n\n  label=type:container_runtime_t \r\n privileged=true ", want: []string{"label=disable", "label=type:container_runtime_t", "privileged=true"}},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, parseTrivySecurityOptsInternal(tt.value))
+		})
+	}
+}
+
+func TestBuildTrivyHostConfig_AppliesRuntimeSecurity(t *testing.T) {
+	hostConfig := buildTrivyHostConfig(
+		"cache-volume",
+		nil,
+		containertypes.Resources{},
+		"",
+		false,
+		"bridge",
+		[]string{"label=disable", "label=type:container_runtime_t"},
+		true,
+	)
+
+	require.Equal(t, []string{"label=disable", "label=type:container_runtime_t"}, hostConfig.SecurityOpt)
+	require.True(t, hostConfig.Privileged)
+}
+
+func TestBuildTrivyBatchHostConfig_AppliesRuntimeSecurity(t *testing.T) {
+	hostConfig := buildTrivyBatchHostConfig(
+		"cache-volume",
+		"arcane-external",
+		[]string{"label=disable"},
+		true,
+	)
+
+	require.Equal(t, containertypes.NetworkMode("arcane-external"), hostConfig.NetworkMode)
+	require.Equal(t, []string{"label=disable"}, hostConfig.SecurityOpt)
+	require.True(t, hostConfig.Privileged)
+	require.Len(t, hostConfig.Mounts, 2)
+}
+
+func TestApplyTrivyRuntimeSecurityInternal(t *testing.T) {
+	hostConfig := &containertypes.HostConfig{}
+
+	applyTrivyRuntimeSecurityInternal(hostConfig, []string{"label=disable"}, true)
+
+	require.Equal(t, []string{"label=disable"}, hostConfig.SecurityOpt)
+	require.True(t, hostConfig.Privileged)
 }
 
 func TestNormalizeTrivyNetworkModeInternal(t *testing.T) {
