+++ frontend/messages/en.json
@@ -975,7 +975,7 @@
 	"security_subtitle": "Aggregated vulnerability findings across all images in this environment.",
 	"security_images_scanned": "Images Scanned",
 	"security_total_vulnerabilities": "Total Vulnerabilities",
-	"security_description": "Configure authentication methods, session timeouts, and other security-related options.",
+	"security_description": "Configure vulnerability scanning and runtime security settings.",
 	"security_authentication_heading": "Authentication",
 	"security_local_auth_label": "Local Authentication",
 	"security_local_auth_description": "Username and password stored by Arcane. Keep enabled as a fallback.",
@@ -1014,6 +1014,13 @@
 	"security_trivy_network_description": "Docker network mode/network name used for Trivy scan containers.",
 	"security_trivy_network_help": "Default is bridge. Use host, none, or a custom Docker network with internet access if needed.",
 	"security_trivy_network_current_value_note": "Current configured value",
+	"security_trivy_security_opts_label": "Trivy Security Options",
+	"security_trivy_security_opts_description": "Docker security options applied to Trivy scan containers.",
+	"security_trivy_security_opts_help": "Use commas or new lines to separate entries. Prefer security options such as label=disable before enabling privileged mode.",
+	"security_trivy_security_opts_placeholder": "label=disable\nlabel=type:container_runtime_t",
+	"security_trivy_privileged_label": "Trivy Privileged",
+	"security_trivy_privileged_description": "Run Trivy scan containers in privileged mode when the host security policy requires it.",
+	"security_trivy_privileged_note": "Use this only as a fallback when security options are not sufficient.",
 	"security_trivy_network_fetch_failed": "Could not load Docker networks; showing built-in options only.",
 	"security_trivy_resource_limits_label": "Trivy Resource Limits",
 	"security_trivy_resource_limits_description": "Configure CPU and memory limits for Trivy scan containers.",
@@ -1031,8 +1038,8 @@
 	"security_settings_saved": "Settings saved successfully.",
 	"security_local_enabled_info": "Local Authentication enabled to ensure at least one provider is active.",
 	"security_settings_save_failed": "Failed to save security settings",
-  	"security_auto_login_enabled_title": "Auto-Login Enabled",
-  	"security_auto_login_enabled_description": "Auto-login is currently enabled via environment variables. Authentication settings are disabled because users are automatically logged in with pre-configured credentials. To modify authentication settings, disable auto-login in your server configuration.",
+	"security_auto_login_enabled_title": "Auto-Login Enabled",
+	"security_auto_login_enabled_description": "Auto-login is currently enabled via environment variables. Authentication settings are disabled because users are automatically logged in with pre-configured credentials. To modify authentication settings, disable auto-login in your server configuration.",
 	"security_scan_all": "Scan all images",
 	"security_scanning": "Scanning",
 	"security_no_images_to_scan": "No images available to scan",
@@ -2187,5 +2194,7 @@
 	"deploy_pull_policy_never": "Never pull",
 	"deploy_force_recreate": "Force recreate containers",
 	"settings_default_deploy_pull_policy": "Default Deploy Pull Policy",
-	"settings_default_deploy_pull_policy_description": "Default image pull policy when deploying projects"
+	"settings_default_deploy_pull_policy_description": "Default image pull policy when deploying projects",
+	"authentication_title": "Authentication",
+	"authentication_description": "Configure authentication providers, password policy, and session behavior."
 }
