+++ frontend/src/lib/config/navigation-config.ts
@@ -4,6 +4,7 @@ import {
 	JobsIcon,
 	UsersIcon,
 	SecurityIcon,
+	LockIcon,
 	NotificationsIcon,
 	DashboardIcon,
 	ProjectsIcon,
@@ -61,15 +62,21 @@ export const navigationItems: Record<string, NavigationItem[]> = {
 				{ title: m.api_key_page_title(), url: '/settings/api-keys', icon: ApiKeyIcon, shortcut: ['mod', 'shift', '1'] },
 				{ title: m.appearance_title(), url: '/settings/appearance', icon: ApperanceIcon, shortcut: ['mod', 'shift', '2'] },
 				{
+					title: m.authentication_title(),
+					url: '/settings/authentication',
+					icon: LockIcon,
+					shortcut: ['mod', 'shift', '3']
+				},
+				{
 					title: m.notifications_title(),
 					url: '/settings/notifications',
 					icon: NotificationsIcon,
-					shortcut: ['mod', 'shift', '3']
+					shortcut: ['mod', 'shift', '4']
 				},
-				{ title: m.security_title(), url: '/settings/security', icon: SecurityIcon, shortcut: ['mod', 'shift', '4'] },
-				{ title: m.builds(), url: '/settings/builds', icon: HammerIcon, shortcut: ['mod', 'shift', '5'] },
-				{ title: m.timeouts_settings(), url: '/settings/timeouts', icon: JobsIcon, shortcut: ['mod', 'shift', '6'] },
-				{ title: m.users_title(), url: '/settings/users', icon: UsersIcon, shortcut: ['mod', 'shift', '7'] }
+				{ title: m.security_title(), url: '/settings/security', icon: SecurityIcon, shortcut: ['mod', 'shift', '5'] },
+				{ title: m.builds(), url: '/settings/builds', icon: HammerIcon, shortcut: ['mod', 'shift', '6'] },
+				{ title: m.timeouts_settings(), url: '/settings/timeouts', icon: JobsIcon, shortcut: ['mod', 'shift', '7'] },
+				{ title: m.users_title(), url: '/settings/users', icon: UsersIcon, shortcut: ['mod', 'shift', '8'] }
 			]
 		}
 	]
