+++ frontend/src/lib/types/settings.type.ts
@@ -40,6 +40,8 @@ export type Settings = {
 	authPasswordPolicy: 'basic' | 'standard' | 'strong';
 	trivyImage: string;
 	trivyNetwork: string;
+	trivySecurityOpts: string;
+	trivyPrivileged: boolean;
 	trivyResourceLimitsEnabled: boolean;
 	trivyCpuLimit: number;
 	trivyMemoryLimitMb: number;
