+++ frontend/src/routes/(app)/settings/security/+page.svelte
@@ -1,23 +1,18 @@
 <script lang="ts">
-	import * as AlertDialog from '$lib/components/ui/alert-dialog';
 	import { z } from 'zod/v4';
 	import { getContext, onMount } from 'svelte';
-	import { ArcaneButton } from '$lib/components/arcane-button/index.js';
-	import { Input } from '$lib/components/ui/input';
 	import { Label } from '$lib/components/ui/label';
 	import { Switch } from '$lib/components/ui/switch/index.js';
-	import { Separator } from '$lib/components/ui/separator';
+	import { Textarea } from '$lib/components/ui/textarea/index.js';
 	import { toast } from 'svelte-sonner';
 	import type { PageData } from './$types';
 	import type { Settings } from '$lib/types/settings.type';
-	import * as ArcaneTooltip from '$lib/components/arcane-tooltip';
 	import { m } from '$lib/paraglide/messages';
-	import { LockIcon, InfoIcon } from '$lib/icons';
+	import { SecurityIcon, InfoIcon } from '$lib/icons';
 	import SearchableSelect from '$lib/components/form/searchable-select.svelte';
 	import TextInputWithLabel from '$lib/components/form/text-input-with-label.svelte';
 	import settingsStore from '$lib/stores/config-store';
 	import { SettingsPageLayout } from '$lib/layouts';
-	import { CopyButton } from '$lib/components/ui/copy-button';
 	import { createSettingsForm } from '$lib/utils/settings-form.util';
 	import * as Alert from '$lib/components/ui/alert';
 	import { networkService } from '$lib/services/network-service';
@@ -25,139 +20,58 @@
 	let { data }: { data: PageData } = $props();
 	const currentSettings = $derived<Settings>($settingsStore || data.settings!);
 	const isReadOnly = $derived.by(() => $settingsStore.uiConfigDisabled);
-	const isAutoLoginEnabled = $derived(settingsStore.autoLoginEnabled.isEnabled());
 
-	const formSchema = z
-		.object({
-			authLocalEnabled: z.boolean(),
-			authSessionTimeout: z.coerce
-				.number()
-				.int(m.security_session_timeout_integer())
-				.min(15, m.security_session_timeout_min())
-				.max(1440, m.security_session_timeout_max()),
-			authPasswordPolicy: z.enum(['basic', 'standard', 'strong']),
-			trivyImage: z.string(),
-			trivyNetwork: z.string(),
-			trivyResourceLimitsEnabled: z.boolean(),
-			trivyCpuLimit: z.coerce.number().int(m.security_session_timeout_integer()).nonnegative(),
-			trivyMemoryLimitMb: z.coerce.number().int().nonnegative(),
-			trivyConcurrentScanContainers: z.coerce.number().int().min(1, m.security_trivy_concurrent_scan_containers_min()),
-			oidcEnabled: z.boolean(),
-			oidcMergeAccounts: z.boolean(),
-			oidcSkipTlsVerify: z.boolean(),
-			oidcAutoRedirectToProvider: z.boolean(),
-			oidcClientId: z.string(),
-			oidcClientSecret: z.string(),
-			oidcIssuerUrl: z.string(),
-			oidcScopes: z.string(),
-			oidcAdminClaim: z.string(),
-			oidcAdminValue: z.string(),
-			oidcProviderName: z.string(),
-			oidcProviderLogoUrl: z.string()
-		})
-		.superRefine((formData, ctx) => {
-			const oidcEnabledForAuthValidation = data.oidcStatus.envForced ? currentSettings.oidcEnabled : formData.oidcEnabled;
-			if (oidcEnabledForAuthValidation) return;
-			if (!formData.authLocalEnabled) {
-				ctx.addIssue({
-					code: 'custom',
-					message: m.security_enable_one_provider(),
-					path: ['authLocalEnabled']
-				});
-			}
-		});
-
-	let showMergeAccountsAlert = $state(false);
+	const formSchema = z.object({
+		trivyImage: z.string(),
+		trivyNetwork: z.string(),
+		trivySecurityOpts: z.string(),
+		trivyPrivileged: z.boolean(),
+		trivyResourceLimitsEnabled: z.boolean(),
+		trivyCpuLimit: z.coerce.number().int(m.security_session_timeout_integer()).nonnegative(),
+		trivyMemoryLimitMb: z.coerce.number().int().nonnegative(),
+		trivyConcurrentScanContainers: z.coerce.number().int().min(1, m.security_trivy_concurrent_scan_containers_min())
+	});
 
 	const formDefaults = $derived({
-		authLocalEnabled: currentSettings.authLocalEnabled,
-		authSessionTimeout: currentSettings.authSessionTimeout,
-		authPasswordPolicy: currentSettings.authPasswordPolicy,
 		trivyImage: currentSettings.trivyImage,
 		trivyNetwork: currentSettings.trivyNetwork || 'bridge',
+		trivySecurityOpts: currentSettings.trivySecurityOpts || '',
+		trivyPrivileged: currentSettings.trivyPrivileged ?? false,
 		trivyResourceLimitsEnabled: currentSettings.trivyResourceLimitsEnabled ?? true,
 		trivyCpuLimit: currentSettings.trivyCpuLimit ?? 1,
 		trivyMemoryLimitMb: currentSettings.trivyMemoryLimitMb ?? 0,
-		trivyConcurrentScanContainers: currentSettings.trivyConcurrentScanContainers ?? 1,
-		oidcEnabled: currentSettings.oidcEnabled,
-		oidcMergeAccounts: currentSettings.oidcMergeAccounts,
-		oidcSkipTlsVerify: currentSettings.oidcSkipTlsVerify,
-		oidcAutoRedirectToProvider: currentSettings.oidcAutoRedirectToProvider,
-		oidcClientId: currentSettings.oidcClientId,
-		oidcClientSecret: '',
-		oidcIssuerUrl: currentSettings.oidcIssuerUrl,
-		oidcScopes: currentSettings.oidcScopes,
-		oidcAdminClaim: currentSettings.oidcAdminClaim,
-		oidcAdminValue: currentSettings.oidcAdminValue,
-		oidcProviderName: currentSettings.oidcProviderName,
-		oidcProviderLogoUrl: currentSettings.oidcProviderLogoUrl
+		trivyConcurrentScanContainers: currentSettings.trivyConcurrentScanContainers ?? 1
 	});
 
-	// Security page needs custom submit logic for OIDC client secret handling
 	let { formInputs, form, settingsForm } = $derived(
 		createSettingsForm({
 			schema: formSchema,
 			currentSettings: formDefaults,
 			getCurrentSettings: () => ({
-				authLocalEnabled: ($settingsStore || data.settings!).authLocalEnabled,
-				authSessionTimeout: ($settingsStore || data.settings!).authSessionTimeout,
-				authPasswordPolicy: ($settingsStore || data.settings!).authPasswordPolicy,
 				trivyImage: ($settingsStore || data.settings!).trivyImage,
 				trivyNetwork: ($settingsStore || data.settings!).trivyNetwork || 'bridge',
+				trivySecurityOpts: ($settingsStore || data.settings!).trivySecurityOpts || '',
+				trivyPrivileged: ($settingsStore || data.settings!).trivyPrivileged ?? false,
 				trivyResourceLimitsEnabled: ($settingsStore || data.settings!).trivyResourceLimitsEnabled ?? true,
 				trivyCpuLimit: ($settingsStore || data.settings!).trivyCpuLimit ?? 1,
 				trivyMemoryLimitMb: ($settingsStore || data.settings!).trivyMemoryLimitMb ?? 0,
-				trivyConcurrentScanContainers: ($settingsStore || data.settings!).trivyConcurrentScanContainers ?? 1,
-				oidcEnabled: ($settingsStore || data.settings!).oidcEnabled,
-				oidcMergeAccounts: ($settingsStore || data.settings!).oidcMergeAccounts,
-				oidcSkipTlsVerify: ($settingsStore || data.settings!).oidcSkipTlsVerify,
-				oidcAutoRedirectToProvider: ($settingsStore || data.settings!).oidcAutoRedirectToProvider,
-				oidcClientId: ($settingsStore || data.settings!).oidcClientId,
-				oidcClientSecret: '',
-				oidcIssuerUrl: ($settingsStore || data.settings!).oidcIssuerUrl,
-				oidcScopes: ($settingsStore || data.settings!).oidcScopes,
-				oidcAdminClaim: ($settingsStore || data.settings!).oidcAdminClaim,
-				oidcAdminValue: ($settingsStore || data.settings!).oidcAdminValue,
-				oidcProviderName: ($settingsStore || data.settings!).oidcProviderName,
-				oidcProviderLogoUrl: ($settingsStore || data.settings!).oidcProviderLogoUrl
+				trivyConcurrentScanContainers: ($settingsStore || data.settings!).trivyConcurrentScanContainers ?? 1
 			}),
 			successMessage: m.security_settings_saved()
 		})
 	);
 
-	// Override the default hasChanges since we need special handling for oidcClientSecret
 	const hasSecurityChanges = $derived(
-		$formInputs.authLocalEnabled.value !== currentSettings.authLocalEnabled ||
-			$formInputs.authSessionTimeout.value !== currentSettings.authSessionTimeout ||
-			$formInputs.authPasswordPolicy.value !== currentSettings.authPasswordPolicy ||
-			$formInputs.trivyImage.value !== currentSettings.trivyImage ||
+		$formInputs.trivyImage.value !== currentSettings.trivyImage ||
 			$formInputs.trivyNetwork.value !== (currentSettings.trivyNetwork || 'bridge') ||
+			$formInputs.trivySecurityOpts.value !== (currentSettings.trivySecurityOpts || '') ||
+			$formInputs.trivyPrivileged.value !== (currentSettings.trivyPrivileged ?? false) ||
 			$formInputs.trivyResourceLimitsEnabled.value !== (currentSettings.trivyResourceLimitsEnabled ?? true) ||
 			$formInputs.trivyCpuLimit.value !== (currentSettings.trivyCpuLimit ?? 1) ||
 			$formInputs.trivyMemoryLimitMb.value !== (currentSettings.trivyMemoryLimitMb ?? 0) ||
-			$formInputs.trivyConcurrentScanContainers.value !== (currentSettings.trivyConcurrentScanContainers ?? 1) ||
-			$formInputs.oidcEnabled.value !== currentSettings.oidcEnabled ||
-			$formInputs.oidcMergeAccounts.value !== currentSettings.oidcMergeAccounts ||
-			$formInputs.oidcSkipTlsVerify.value !== currentSettings.oidcSkipTlsVerify ||
-			$formInputs.oidcAutoRedirectToProvider.value !== currentSettings.oidcAutoRedirectToProvider ||
-			$formInputs.oidcClientId.value !== currentSettings.oidcClientId ||
-			$formInputs.oidcIssuerUrl.value !== currentSettings.oidcIssuerUrl ||
-			$formInputs.oidcScopes.value !== currentSettings.oidcScopes ||
-			$formInputs.oidcAdminClaim.value !== currentSettings.oidcAdminClaim ||
-			$formInputs.oidcAdminValue.value !== currentSettings.oidcAdminValue ||
-			$formInputs.oidcProviderName.value !== currentSettings.oidcProviderName ||
-			$formInputs.oidcProviderLogoUrl.value !== currentSettings.oidcProviderLogoUrl ||
-			$formInputs.oidcClientSecret.value !== ''
+			$formInputs.trivyConcurrentScanContainers.value !== (currentSettings.trivyConcurrentScanContainers ?? 1)
 	);
 
-	const redirectUri = $derived(`${globalThis?.location?.origin ?? ''}/auth/oidc/callback`);
-	const isOidcEnvForced = $derived(data.oidcStatus.envForced);
-	const isOidcForcedEnabled = $derived(isOidcEnvForced && currentSettings.oidcEnabled);
-	const isOidcForcedDisabled = $derived(isOidcEnvForced && !currentSettings.oidcEnabled);
-	const isOidcEnabledForAuthValidation = $derived.by(() =>
-		isOidcEnvForced ? currentSettings.oidcEnabled : $formInputs.oidcEnabled.value
-	);
-	const showOidcDetails = $derived($formInputs.oidcEnabled.value || isOidcForcedEnabled);
 	const baseTrivyNetworkOptions = [
 		{ value: 'bridge', label: 'bridge' },
 		{ value: 'host', label: 'host' },
@@ -227,40 +141,19 @@
 		const trivyCpuLimit = formData.trivyResourceLimitsEnabled ? formData.trivyCpuLimit : 0;
 		const trivyMemoryLimitMb = formData.trivyResourceLimitsEnabled ? formData.trivyMemoryLimitMb : 0;
 
-		if (formData.oidcEnabled && !isOidcEnvForced) {
-			if (!formData.oidcClientId || !formData.oidcIssuerUrl) {
-				toast.error(m.security_oidc_required_fields());
-				return;
-			}
-		}
-
 		settingsForm.setLoading(true);
 
 		try {
 			await settingsForm.updateSettings({
-				authLocalEnabled: formData.authLocalEnabled,
-				authSessionTimeout: formData.authSessionTimeout,
-				authPasswordPolicy: formData.authPasswordPolicy,
 				trivyImage: formData.trivyImage,
 				trivyNetwork: formData.trivyNetwork,
+				trivySecurityOpts: formData.trivySecurityOpts,
+				trivyPrivileged: formData.trivyPrivileged,
 				trivyResourceLimitsEnabled: formData.trivyResourceLimitsEnabled,
 				trivyCpuLimit,
 				trivyMemoryLimitMb,
-				trivyConcurrentScanContainers: formData.trivyConcurrentScanContainers,
-				oidcEnabled: formData.oidcEnabled,
-				oidcMergeAccounts: formData.oidcMergeAccounts,
-				oidcSkipTlsVerify: formData.oidcSkipTlsVerify,
-				oidcAutoRedirectToProvider: formData.oidcAutoRedirectToProvider,
-				oidcClientId: formData.oidcClientId,
-				oidcIssuerUrl: formData.oidcIssuerUrl,
-				oidcScopes: formData.oidcScopes,
-				oidcAdminClaim: formData.oidcAdminClaim,
-				oidcAdminValue: formData.oidcAdminValue,
-				oidcProviderName: formData.oidcProviderName,
-				oidcProviderLogoUrl: formData.oidcProviderLogoUrl,
-				...(formData.oidcClientSecret && { oidcClientSecret: formData.oidcClientSecret })
+				trivyConcurrentScanContainers: formData.trivyConcurrentScanContainers
 			});
-			$formInputs.oidcClientSecret.value = '';
 			toast.success(m.security_settings_saved());
 		} catch (error: any) {
 			console.error('Failed to save settings:', error);
@@ -272,32 +165,6 @@
 
 	function customReset() {
 		form.reset();
-		$formInputs.oidcClientSecret.value = '';
-	}
-
-	function handleLocalSwitchChange(checked: boolean) {
-		if (!checked && !isOidcEnabledForAuthValidation) {
-			$formInputs.authLocalEnabled.value = true;
-			toast.error(m.security_enable_one_provider_error());
-			return;
-		}
-		$formInputs.authLocalEnabled.value = checked;
-	}
-
-	function handleOidcEnabledChange(checked: boolean) {
-		if (!checked && !$formInputs.authLocalEnabled.value && !isOidcEnvForced) {
-			$formInputs.authLocalEnabled.value = true;
-			toast.info(m.security_local_enabled_info());
-		}
-		$formInputs.oidcEnabled.value = checked;
-	}
-
-	function handleMergeAccountsChange(checked: boolean) {
-		if (checked && !currentSettings.oidcMergeAccounts) {
-			showMergeAccountsAlert = true;
-		} else {
-			$formInputs.oidcMergeAccounts.value = checked;
-		}
 	}
 
 	function handleTrivyResourceLimitsChange(checked: boolean) {
@@ -308,24 +175,12 @@
 		}
 	}
 
-	function confirmMergeAccounts() {
-		$formInputs.oidcMergeAccounts.value = true;
-		showMergeAccountsAlert = false;
-	}
-
-	function cancelMergeAccounts() {
-		$formInputs.oidcMergeAccounts.value = false;
-		showMergeAccountsAlert = false;
-	}
-
 	onMount(() => {
 		void loadTrivyNetworkOptions();
 	});
 
 	$effect(() => {
-		// Use custom submit/reset for security page
 		settingsForm.registerFormActions(customSubmit, customReset);
-		// Sync the custom hasSecurityChanges to the context
 		const formState = getContext('settingsFormState') as any;
 		if (formState) {
 			formState.hasChanges = hasSecurityChanges;
@@ -336,388 +191,12 @@
 <SettingsPageLayout
 	title={m.security_title()}
 	description={m.security_description()}
-	icon={LockIcon}
+	icon={SecurityIcon}
 	pageType="form"
 	showReadOnlyTag={isReadOnly}
 >
 	{#snippet mainContent()}
 		<fieldset disabled={isReadOnly} class="relative space-y-8">
-			<!-- Authentication Section -->
-			<div class="space-y-4">
-				<h3 class="text-lg font-medium">{m.security_authentication_heading()}</h3>
-
-				{#if isAutoLoginEnabled}
-					<!-- Auto-Login Enabled Alert -->
-					<Alert.Root variant="default" class="border-amber-200 bg-amber-50 dark:border-amber-800 dark:bg-amber-950">
-						<InfoIcon class="h-4 w-4 text-amber-600 dark:text-amber-500" />
-						<Alert.Title class="text-amber-900 dark:text-amber-100">{m.security_auto_login_enabled_title()}</Alert.Title>
-						<Alert.Description class="text-amber-800 dark:text-amber-200">
-							{m.security_auto_login_enabled_description()}
-						</Alert.Description>
-					</Alert.Root>
-				{:else}
-					<!-- Authentication Cards (only shown when auto-login is disabled) -->
-					<div class="bg-card rounded-lg border shadow-sm">
-						<div class="space-y-6 p-6">
-							<!-- Local Auth -->
-							<div class="grid gap-4 md:grid-cols-[1fr_1.5fr] md:gap-8">
-								<div>
-									<Label class="text-base">{m.security_local_auth_label()}</Label>
-									<p class="text-muted-foreground mt-1 text-sm">{m.security_local_auth_description()}</p>
-								</div>
-								<div class="flex items-center gap-2">
-									<Switch
-										id="localAuthSwitch"
-										bind:checked={$formInputs.authLocalEnabled.value}
-										onCheckedChange={handleLocalSwitchChange}
-									/>
-									<Label for="localAuthSwitch" class="font-normal">
-										{$formInputs.authLocalEnabled.value ? m.common_enabled() : m.common_disabled()}
-									</Label>
-								</div>
-							</div>
-
-							<Separator />
-
-							<!-- OIDC Auth -->
-							<div class="grid gap-4 md:grid-cols-[1fr_1.5fr] md:gap-8">
-								<div>
-									<Label class="text-base">{m.security_oidc_auth_label()}</Label>
-									<p class="text-muted-foreground mt-1 text-sm">{m.security_oidc_auth_description()}</p>
-									{#if isOidcEnvForced}
-										<div class="mt-2">
-											<ArcaneTooltip.Root>
-												<ArcaneTooltip.Trigger>
-													<span
-														class="inline-flex items-center gap-1.5 rounded-full bg-amber-100 px-2.5 py-1 text-xs font-medium text-amber-800 ring-1 ring-amber-200 dark:bg-amber-900/50 dark:text-amber-200 dark:ring-amber-800"
-													>
-														{#if isOidcForcedDisabled}
-															{m.security_server_disabled_via_server()}
-														{:else}
-															{m.security_server_configured()}
-														{/if}
-													</span>
-												</ArcaneTooltip.Trigger>
-												<ArcaneTooltip.Content side="top">
-													{#if isOidcForcedDisabled}
-														{m.security_oidc_forced_disabled_tooltip()}
-													{:else}
-														{m.security_oidc_forced_managed_tooltip()}
-													{/if}
-												</ArcaneTooltip.Content>
-											</ArcaneTooltip.Root>
-										</div>
-									{/if}
-								</div>
-								<div class="space-y-4">
-									<div class="flex items-center gap-2">
-										<Switch
-											id="oidcEnabledSwitch"
-											disabled={isOidcEnvForced}
-											bind:checked={$formInputs.oidcEnabled.value}
-											onCheckedChange={handleOidcEnabledChange}
-										/>
-										<Label for="oidcEnabledSwitch" class="font-normal">
-											{m.security_oidc_enabled_label()}
-										</Label>
-									</div>
-
-									{#if showOidcDetails}
-										<div class="space-y-4 pt-2">
-											<div class="space-y-2">
-												<Label for="oidcClientId" class="text-sm font-medium">{m.oidc_client_id_label()}</Label>
-												<Input
-													id="oidcClientId"
-													type="text"
-													placeholder={m.oidc_client_id_placeholder()}
-													disabled={isOidcEnvForced}
-													bind:value={$formInputs.oidcClientId.value}
-													class="font-mono text-sm"
-												/>
-												{#if $formInputs.oidcClientId.error}
-													<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcClientId.error}</p>
-												{/if}
-											</div>
-
-											<div class="space-y-2">
-												<Label for="oidcClientSecret" class="text-sm font-medium">{m.oidc_client_secret_label()}</Label>
-												<Input
-													id="oidcClientSecret"
-													type="password"
-													placeholder={m.oidc_client_secret_placeholder()}
-													disabled={isOidcEnvForced}
-													bind:value={$formInputs.oidcClientSecret.value}
-													class="font-mono text-sm"
-												/>
-												<p class="text-muted-foreground text-xs">{m.security_oidc_client_secret_help()}</p>
-												{#if $formInputs.oidcClientSecret.error}
-													<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcClientSecret.error}</p>
-												{/if}
-											</div>
-
-											<div class="space-y-2">
-												<Label for="oidcIssuerUrl" class="text-sm font-medium">{m.oidc_issuer_url_label()}</Label>
-												<Input
-													id="oidcIssuerUrl"
-													type="text"
-													placeholder={m.oidc_issuer_url_placeholder()}
-													disabled={isOidcEnvForced}
-													bind:value={$formInputs.oidcIssuerUrl.value}
-													class="font-mono text-sm"
-												/>
-												<p class="text-muted-foreground text-xs">{m.oidc_issuer_url_description()}</p>
-												{#if $formInputs.oidcIssuerUrl.error}
-													<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcIssuerUrl.error}</p>
-												{/if}
-											</div>
-
-											<div class="space-y-2">
-												<Label for="oidcProviderName" class="text-sm font-medium">{m.oidc_provider_name_label()}</Label>
-												<Input
-													id="oidcProviderName"
-													type="text"
-													placeholder={m.oidc_provider_name_placeholder()}
-													disabled={isOidcEnvForced}
-													bind:value={$formInputs.oidcProviderName.value}
-													class="font-mono text-sm"
-												/>
-												<p class="text-muted-foreground text-xs">{m.oidc_provider_name_description()}</p>
-												{#if $formInputs.oidcProviderName.error}
-													<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcProviderName.error}</p>
-												{/if}
-											</div>
-
-											<div class="space-y-2">
-												<Label for="oidcProviderLogoUrl" class="text-sm font-medium">{m.oidc_provider_logo_url_label()}</Label>
-												<Input
-													id="oidcProviderLogoUrl"
-													type="text"
-													placeholder={m.oidc_provider_logo_url_placeholder()}
-													disabled={isOidcEnvForced}
-													bind:value={$formInputs.oidcProviderLogoUrl.value}
-													class="font-mono text-sm"
-												/>
-												<p class="text-muted-foreground text-xs">{m.oidc_provider_logo_url_description()}</p>
-												{#if $formInputs.oidcProviderLogoUrl.error}
-													<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcProviderLogoUrl.error}</p>
-												{/if}
-											</div>
-
-											<div class="space-y-2">
-												<Label for="oidcScopes" class="text-sm font-medium">{m.oidc_scopes_label()}</Label>
-												<Input
-													id="oidcScopes"
-													type="text"
-													placeholder={m.oidc_scopes_placeholder()}
-													disabled={isOidcEnvForced}
-													bind:value={$formInputs.oidcScopes.value}
-													class="font-mono text-sm"
-												/>
-												{#if $formInputs.oidcScopes.error}
-													<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcScopes.error}</p>
-												{/if}
-											</div>
-
-											<div class="border-t pt-4">
-												<h4 class="text-sm font-semibold">{m.oidc_admin_role_mapping_title()}</h4>
-												<p class="text-muted-foreground mb-3 text-xs">{m.oidc_admin_role_mapping_description()}</p>
-												<div class="grid gap-3 sm:grid-cols-2">
-													<div class="space-y-2">
-														<Label for="oidcAdminClaim" class="text-sm font-medium">{m.oidc_admin_claim_label()}</Label>
-														<Input
-															id="oidcAdminClaim"
-															type="text"
-															placeholder={m.oidc_admin_claim_placeholder()}
-															disabled={isOidcEnvForced}
-															bind:value={$formInputs.oidcAdminClaim.value}
-															class="font-mono text-sm"
-														/>
-														{#if $formInputs.oidcAdminClaim.error}
-															<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcAdminClaim.error}</p>
-														{/if}
-													</div>
-													<div class="space-y-2">
-														<Label for="oidcAdminValue" class="text-sm font-medium">{m.oidc_admin_value_label()}</Label>
-														<Input
-															id="oidcAdminValue"
-															type="text"
-															placeholder={m.oidc_admin_value_placeholder()}
-															disabled={isOidcEnvForced}
-															bind:value={$formInputs.oidcAdminValue.value}
-															class="font-mono text-sm"
-														/>
-														<p class="text-muted-foreground text-[11px]">{m.oidc_admin_value_help()}</p>
-														{#if $formInputs.oidcAdminValue.error}
-															<p class="text-destructive text-[0.8rem] font-medium">{$formInputs.oidcAdminValue.error}</p>
-														{/if}
-													</div>
-												</div>
-											</div>
-
-											<div class="border-t pt-4">
-												<div class="flex items-center gap-2">
-													<Switch
-														id="oidcMergeAccountsSwitch"
-														disabled={isOidcEnvForced}
-														bind:checked={$formInputs.oidcMergeAccounts.value}
-														onCheckedChange={handleMergeAccountsChange}
-													/>
-													<div class="grid gap-1.5 leading-none">
-														<Label for="oidcMergeAccountsSwitch" class="font-normal">
-															{m.security_oidc_merge_accounts_label()}
-														</Label>
-														<p class="text-muted-foreground text-xs">
-															{m.security_oidc_merge_accounts_description()}
-														</p>
-													</div>
-												</div>
-											</div>
-
-											<div class="border-t pt-4">
-												<div class="flex items-center gap-2">
-													<Switch
-														id="oidcSkipTlsVerifySwitch"
-														disabled={isOidcEnvForced}
-														bind:checked={$formInputs.oidcSkipTlsVerify.value}
-													/>
-													<div class="grid gap-1.5 leading-none">
-														<Label for="oidcSkipTlsVerifySwitch" class="font-normal">
-															{m.oidc_skip_tls_verify_label()}
-														</Label>
-														<p class="text-muted-foreground text-xs">
-															{m.oidc_skip_tls_verify_description()}
-														</p>
-													</div>
-												</div>
-											</div>
-
-											<div class="border-t pt-4">
-												<div class="flex items-center gap-2">
-													<Switch
-														id="oidcAutoRedirectSwitch"
-														disabled={isOidcEnvForced}
-														bind:checked={$formInputs.oidcAutoRedirectToProvider.value}
-													/>
-													<div class="grid gap-1.5 leading-none">
-														<Label for="oidcAutoRedirectSwitch" class="font-normal">
-															{m.oidc_auto_redirect_label()}
-														</Label>
-														<p class="text-muted-foreground text-xs">
-															{m.oidc_auto_redirect_description()}
-														</p>
-													</div>
-												</div>
-											</div>
-
-											<div class="bg-muted/30 rounded-lg border p-4">
-												<div class="mb-2 flex items-center gap-2">
-													<InfoIcon class="size-4 text-blue-600" />
-													<span class="text-sm font-medium">{m.oidc_redirect_uri_title()}</span>
-												</div>
-												<p class="text-muted-foreground mb-3 text-sm">{m.oidc_redirect_uri_description()}</p>
-												<div class="flex items-center gap-2">
-													<code class="bg-muted flex-1 rounded p-2 font-mono text-xs break-all">{redirectUri}</code>
-													<CopyButton text={redirectUri} size="sm" variant="outline" class="shrink-0" title={m.common_copy()} />
-												</div>
-											</div>
-										</div>
-									{/if}
-								</div>
-							</div>
-						</div>
-					</div>
-				{/if}
-			</div>
-
-			<!-- Session Section -->
-			<div class="space-y-4">
-				<h3 class="text-lg font-medium">{m.security_session_heading()}</h3>
-				<div class="bg-card rounded-lg border shadow-sm">
-					<div class="space-y-6 p-6">
-						<div class="grid gap-4 md:grid-cols-[1fr_1.5fr] md:gap-8">
-							<div>
-								<Label class="text-base">{m.security_session_timeout_label()}</Label>
-								<p class="text-muted-foreground mt-1 text-sm">{m.security_session_timeout_description()}</p>
-							</div>
-							<div class="max-w-xs">
-								<TextInputWithLabel
-									bind:value={$formInputs.authSessionTimeout.value}
-									error={$formInputs.authSessionTimeout.error}
-									label={m.security_session_timeout_label()}
-									placeholder={m.security_session_timeout_placeholder()}
-									helpText={m.security_session_timeout_description()}
-									type="number"
-								/>
-							</div>
-						</div>
-					</div>
-				</div>
-			</div>
-
-			<!-- Password Policy Section -->
-			<div class="space-y-4">
-				<h3 class="text-lg font-medium">{m.security_password_policy_label()}</h3>
-				<div class="bg-card rounded-lg border shadow-sm">
-					<div class="space-y-6 p-6">
-						<div class="grid gap-4 md:grid-cols-[1fr_1.5fr] md:gap-8">
-							<div>
-								<Label class="text-base">{m.security_password_policy_label()}</Label>
-								<p class="text-muted-foreground mt-1 text-sm">{m.security_password_policy_description()}</p>
-							</div>
-							<div>
-								<div class="grid grid-cols-1 gap-2 sm:grid-cols-3 sm:gap-3" role="group" aria-labelledby="passwordPolicyLabel">
-									<ArcaneTooltip.Root>
-										<ArcaneTooltip.Trigger>
-											<ArcaneButton
-												action="base"
-												tone={$formInputs.authPasswordPolicy.value === 'basic' ? 'outline-primary' : 'outline'}
-												class="h-12 w-full text-xs sm:text-sm"
-												onclick={() => ($formInputs.authPasswordPolicy.value = 'basic')}
-												customLabel={m.common_basic()}
-											/>
-										</ArcaneTooltip.Trigger>
-										<ArcaneTooltip.Content side="top">
-											{m.security_password_policy_basic_tooltip()}
-										</ArcaneTooltip.Content>
-									</ArcaneTooltip.Root>
-
-									<ArcaneTooltip.Root>
-										<ArcaneTooltip.Trigger>
-											<ArcaneButton
-												action="base"
-												tone={$formInputs.authPasswordPolicy.value === 'standard' ? 'outline-primary' : 'outline'}
-												class="h-12 w-full text-xs sm:text-sm"
-												onclick={() => ($formInputs.authPasswordPolicy.value = 'standard')}
-												customLabel={m.security_password_policy_standard()}
-											/>
-										</ArcaneTooltip.Trigger>
-										<ArcaneTooltip.Content side="top">
-											{m.security_password_policy_standard_tooltip()}
-										</ArcaneTooltip.Content>
-									</ArcaneTooltip.Root>
-
-									<ArcaneTooltip.Root>
-										<ArcaneTooltip.Trigger>
-											<ArcaneButton
-												action="base"
-												tone={$formInputs.authPasswordPolicy.value === 'strong' ? 'outline-primary' : 'outline'}
-												class="h-12 w-full text-xs sm:text-sm"
-												onclick={() => ($formInputs.authPasswordPolicy.value = 'strong')}
-												customLabel={m.security_password_policy_strong()}
-											/>
-										</ArcaneTooltip.Trigger>
-										<ArcaneTooltip.Content side="top">
-											{m.security_password_policy_strong_tooltip()}
-										</ArcaneTooltip.Content>
-									</ArcaneTooltip.Root>
-								</div>
-							</div>
-						</div>
-					</div>
-				</div>
-			</div>
-
-			<!-- Vulnerability Scanning Section -->
 			<div class="space-y-4">
 				<h3 class="text-lg font-medium">{m.security_vulnerability_scanning_heading()}</h3>
 				<div class="bg-card rounded-lg border shadow-sm">
@@ -759,6 +238,50 @@
 								/>
 								{#if $formInputs.trivyNetwork.error}
 									<p class="text-destructive mt-2 text-sm">{$formInputs.trivyNetwork.error}</p>
+								{/if}
+							</div>
+						</div>
+
+						<div class="grid gap-4 md:grid-cols-[1fr_1.5fr] md:gap-8">
+							<div>
+								<Label class="text-base">{m.security_trivy_security_opts_label()}</Label>
+								<p class="text-muted-foreground mt-1 text-sm">{m.security_trivy_security_opts_description()}</p>
+								<p class="text-muted-foreground mt-2 text-xs">{m.security_trivy_security_opts_help()}</p>
+							</div>
+							<div class="space-y-2">
+								<Textarea
+									bind:value={$formInputs.trivySecurityOpts.value}
+									aria-label={m.security_trivy_security_opts_label()}
+									class="min-h-28 font-mono text-sm"
+									placeholder={m.security_trivy_security_opts_placeholder()}
+									rows={4}
+								/>
+								{#if $formInputs.trivySecurityOpts.error}
+									<p class="text-destructive text-sm">{$formInputs.trivySecurityOpts.error}</p>
+								{/if}
+							</div>
+						</div>
+
+						<div class="grid gap-4 md:grid-cols-[1fr_1.5fr] md:gap-8">
+							<div>
+								<Label class="text-base">{m.security_trivy_privileged_label()}</Label>
+								<p class="text-muted-foreground mt-1 text-sm">{m.security_trivy_privileged_description()}</p>
+								<p class="text-muted-foreground mt-2 text-xs">{m.security_trivy_privileged_note()}</p>
+							</div>
+							<div class="space-y-3">
+								<div class="flex items-center gap-2">
+									<Switch id="trivyPrivilegedSwitch" bind:checked={$formInputs.trivyPrivileged.value} />
+									<Label for="trivyPrivilegedSwitch" class="font-normal">
+										{$formInputs.trivyPrivileged.value ? m.common_enabled() : m.common_disabled()}
+									</Label>
+								</div>
+								{#if $formInputs.trivyPrivileged.value}
+									<Alert.Root variant="default" class="border-amber-200 bg-amber-50 dark:border-amber-800 dark:bg-amber-950">
+										<InfoIcon class="h-4 w-4 text-amber-900 dark:text-amber-100" />
+										<Alert.Description class="text-amber-800 dark:text-amber-200">
+											{m.security_trivy_privileged_note()}
+										</Alert.Description>
+									</Alert.Root>
 								{/if}
 							</div>
 						</div>
@@ -813,21 +336,5 @@
 				</div>
 			</div>
 		</fieldset>
-	{/snippet}
-	{#snippet additionalContent()}
-		<AlertDialog.Root bind:open={showMergeAccountsAlert}>
-			<AlertDialog.Content>
-				<AlertDialog.Header>
-					<AlertDialog.Title>{m.security_oidc_merge_accounts_alert_title()}</AlertDialog.Title>
-					<AlertDialog.Description>
-						{m.security_oidc_merge_accounts_alert_description()}
-					</AlertDialog.Description>
-				</AlertDialog.Header>
-				<AlertDialog.Footer>
-					<AlertDialog.Cancel onclick={cancelMergeAccounts}>{m.common_cancel()}</AlertDialog.Cancel>
-					<AlertDialog.Action onclick={confirmMergeAccounts}>{m.common_confirm()}</AlertDialog.Action>
-				</AlertDialog.Footer>
-			</AlertDialog.Content>
-		</AlertDialog.Root>
 	{/snippet}
 </SettingsPageLayout>
